--- test/trtest2.js
@@ -1,12 +1,0 @@
-//
-// Integration tests in a telerivet environment
-// (Second block)
-//
-
-var assert = require('./assert');
-
-var api = require('../api');
-
-assert.ok(state.vars[api.persistVar]);
-assert.equal(api.url, "URL");
-assert.equal(api.key, "APIKEY");
